Updating Metal version and removing data type overrides from MNIST model

+++ pybuda/test/mlir/mnist/utils.py
@@ -16,11 +16,11 @@ from torchvision.datasets import MNIST as mnist_dataset
 class MNISTLinear(nn.Module):
     def __init__(self, input_size=784, output_size=10, hidden_size=256):
         super(MNISTLinear, self).__init__()
-        self.l1 = nn.Linear(input_size, hidden_size, bias=False, dtype=torch.bfloat16)
-        self.b1 = nn.Parameter(torch.ones(1, hidden_size, dtype=torch.bfloat16))
+        self.l1 = nn.Linear(input_size, hidden_size, bias=False)
+        self.b1 = nn.Parameter(torch.ones(1, hidden_size))
         self.relu = nn.ReLU()
-        self.b2 = nn.Parameter(torch.ones(1, output_size, dtype=torch.bfloat16))
-        self.l2 = nn.Linear(hidden_size, output_size, bias=False, dtype=torch.bfloat16)
+        self.b2 = nn.Parameter(torch.ones(1, output_size))
+        self.l2 = nn.Linear(hidden_size, output_size, bias=False)
 
     def forward(self, x):
         x = self.l1(x)
@@ -29,7 +29,7 @@ class MNISTLinear(nn.Module):
         x = self.l2(x)
         x = x + self.b2
 
-        return nn.functional.softmax(x, dtype=torch.bfloat16)
+        return nn.functional.softmax(x)
